requirements.txt
lxml==5.3.0
PyPDF2==3.0.1
python-docx==1.1.2
typing_extensions==4.12.2
